services/search/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
httptools==0.6.4
uvloop==0.21.0; sys_platform != 'win32'
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.9.1
sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
numpy>=1.26.0
